pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mccode-antlr"
dependencies = [
    'antlr4-tools==0.2',
    'antlr4-python3-runtime==4.13.0',
    'numpy',
    'pooch',
    'confuse==2.0.1',
    'zenlog==1.1,',
    "importlib_metadata; python_version<'3.8'",
]
dynamic = ["version"]

[project.optional-dependencies]
test = ["gputil"]

[project.scripts]
mcstas-antlr = "mccode_antlr.commands:mcstas"
mcxtrace-antlr = "mccode_antlr.commands:mcxtrace"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
where = ["."]
include = ["mccode_antlr*"]